Added fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,17 +9,16 @@
 
 int main(void)
 {
-	int count, new = 1, prev = 1, temp;
+	long count, new = 1, prev = 1, temp;
 
-	printf("%d, ",prev);
-	for(count = 0; count < 50; count++)
+	printf("%d, ", prev);
+	for (count = 0; count < 50; count++)
 	{
 		temp = new;
 		new += prev;
 		printf("%d, ", new);
 
 		prev = temp;
-	}	
-
+	}
 	return (0);
 }
